Align all check to return boolean and move method in ts
